backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.4

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.1
python-decouple==3.8
django-allauth==0.57.0
dj-rest-auth==5.0.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Payments
stripe==7.0.0
paypalrestsdk==1.13.1

# Email & Notifications
django-anymail==10.0

# API & Serialization
requests==2.31.0
python-dateutil==2.8.2

# File Handling
Pillow==10.1.0
django-storages==1.14.2

# AWS S3 (optional)
boto3==1.32.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
pylint==3.0.3

# Documentation
drf-spectacular==0.27.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.38.0

# Utilities
python-slugify==8.0.1
django-model-utils==4.3.1
